CORRECCION DE ERROR MOSTRAR NOMBRE Y PRECIO
